requirements.txt
anilistpy==0.0.5.2
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
decorating==0.6.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.4
entrypoints==0.3
idna==2.10
importlib-metadata==3.3.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.2
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
mal-api==0.3.7
MarkupSafe==1.1.1
mistune==0.8.4
mock==3.0.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.6
numpy==1.20.3
opencv-python==4.5.2.52
packaging==20.8
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.9.0
prompt-toolkit==3.0.8
pycparser==2.20
Pygments==2.7.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
pywin32==300
pywinpty==0.5.7
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.1
requirements-builder==0.4.2
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
terminado==0.9.1
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.0
